bulk-cdk-core-extract: replace AirbyteStreamDecorator with AirbyteStreamFactory
