+++ airbyte-cdk/bulk/core/extract/src/main/kotlin/io/airbyte/cdk/discover/DiscoveredStream.kt
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) 2024 Airbyte, Inc., all rights reserved.
+ */
+
+package io.airbyte.cdk.discover
+
+data class DiscoveredStream(
+    val name: String,
+    val namespace: String?,
+    val columns: List<Field>,
+    val primaryKeyColumnIDs: List<List<String>>,
+)
