+++ airbyte-cdk/bulk/core/extract/src/test/resources/fakesource/cdc-catalog.json
@@ -21,7 +21,8 @@
           }
         },
         "supported_sync_modes": ["full_refresh", "incremental"],
-        "default_cursor_field": ["ID", "TS"],
+        "source_defined_cursor": false,
+        "default_cursor_field": [],
         "source_defined_primary_key": [["ID"]],
         "is_resumable": true,
         "namespace": "PUBLIC"
@@ -47,7 +48,8 @@
           }
         },
         "supported_sync_modes": ["full_refresh", "incremental"],
-        "default_cursor_field": ["K"],
+        "source_defined_cursor": false,
+        "default_cursor_field": [],
         "source_defined_primary_key": [["K"]],
         "is_resumable": true,
         "namespace": "PUBLIC"
