+++ airbyte-cdk/bulk/core/extract/src/test/resources/fakesource/expected-cursor-catalog.json
@@ -24,6 +24,7 @@
         }
       },
       "supported_sync_modes": ["full_refresh", "incremental"],
+      "source_defined_cursor": false,
       "default_cursor_field": [],
       "source_defined_primary_key": [["ID"]],
       "is_resumable": true,
@@ -48,6 +49,7 @@
         }
       },
       "supported_sync_modes": ["full_refresh", "incremental"],
+      "source_defined_cursor": false,
       "default_cursor_field": [],
       "source_defined_primary_key": [["K"]],
       "is_resumable": true,
