+++ airbyte-cdk/bulk/core/extract/src/test/resources/read/cdc-catalog.json
@@ -20,7 +20,8 @@
           }
         },
         "supported_sync_modes": ["full_refresh", "incremental"],
-        "default_cursor_field": ["ID", "TS"],
+        "source_defined_cursor": false,
+        "default_cursor_field": [],
         "source_defined_primary_key": [["ID"]],
         "is_resumable": true,
         "namespace": "PUBLIC"
@@ -46,7 +47,8 @@
           }
         },
         "supported_sync_modes": ["full_refresh", "incremental"],
-        "default_cursor_field": ["K"],
+        "source_defined_cursor": false,
+        "default_cursor_field": [],
         "source_defined_primary_key": [["K"]],
         "is_resumable": true,
         "namespace": "PUBLIC"
