+++ airbyte-cdk/bulk/core/extract/src/test/resources/read/cursor-catalog.json
@@ -20,6 +20,7 @@
           }
         },
         "supported_sync_modes": ["full_refresh", "incremental"],
+        "source_defined_cursor": false,
         "default_cursor_field": [],
         "source_defined_primary_key": [["ID"]],
         "is_resumable": true,
@@ -46,6 +47,7 @@
           }
         },
         "supported_sync_modes": ["full_refresh", "incremental"],
+        "source_defined_cursor": false,
         "default_cursor_field": [],
         "source_defined_primary_key": [["K"]],
         "is_resumable": true,
